+++ CHANGELOG.md
@@ -41,6 +41,7 @@ animation bug in multi-player and occasionally sticking to the ground in some le
 - fix debris object pool (used for ejected shells) easly getting exhausted when local player is dead
 - fix loading images with color palette in 32-bit mode
 - fix default player weapon being incorrect in listen server after playing on a modded server
+- fix cutscene cinema bars when MSAA is enabled (regression)
 
 Version 1.3.0 (released 2019-05-07)
 -----------------------------------